added: support for backfillTickRate < 1Hz

+++ modules/core/src/main/scala/flashbot/core/DataSource.scala
@@ -74,7 +74,7 @@ class DataSource {
     */
   def types: Seq[DataType[_]] = Seq.empty
 
-  protected[flashbot] def backfillTickRate: Int = 1
+  protected[flashbot] def backfillTickRate: Double = 1d
 }
 
 object DataSource {
